data #1 : User Type을UserInfo와 UserData로 분리
