+++ Outline/Outline/Source/Model/CourseModel.swift
@@ -1,0 +1,18 @@
+//
+//  CourseModel.swift
+//  Outline
+//
+//  Created by Seungui Moon on 10/15/23.
+//
+
+import Foundation
+
+protocol CourseModelProtocol {
+    
+}
+
+struct CourseModel: CourseModelProtocol {
+    
+    func readAllCourses(){}
+    func readCourse(id: String){}
+}
